Handle delete gracefully and corresponding tests

+++ src/main/java/com/ecommerce/electronicsstore/controller/ProductController.java
@@ -47,8 +47,14 @@ public class ProductController {
     @DeleteMapping("/{id}")
   
     public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
-        productService.deleteProduct(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        Product product = productService.getProductById(id);
+        if (product != null) {
+            productService.deleteProduct(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
     }
 
 }
